Split Git and GitHub client helpers
