+++ supacode/Clients/Github/GithubCLIClient.swift
@@ -2,72 +2,6 @@ import ComposableArchitecture
 import Darwin
 import Foundation
 
-struct GithubAuthStatus: Equatable, Sendable {
-  let username: String
-  let host: String
-}
-
-private struct GithubAuthStatusResponse: Sendable {
-  let hosts: [String: [GithubAuthAccount]]
-
-  struct GithubAuthAccount: Sendable {
-    let active: Bool
-    let login: String
-  }
-}
-
-extension GithubAuthStatusResponse: Decodable {
-  private enum CodingKeys: String, CodingKey {
-    case hosts
-  }
-
-  nonisolated init(from decoder: any Decoder) throws {
-    let container = try decoder.container(keyedBy: CodingKeys.self)
-    self.hosts = try container.decode([String: [GithubAuthAccount]].self, forKey: .hosts)
-  }
-}
-
-extension GithubAuthStatusResponse.GithubAuthAccount: Decodable {
-  private enum CodingKeys: String, CodingKey {
-    case active
-    case login
-  }
-
-  nonisolated init(from decoder: any Decoder) throws {
-    let container = try decoder.container(keyedBy: CodingKeys.self)
-    self.active = try container.decode(Bool.self, forKey: .active)
-    self.login = try container.decode(String.self, forKey: .login)
-  }
-}
-
-nonisolated struct RepoKey: Hashable, Sendable {
-  let owner: String
-  let repo: String
-}
-
-nonisolated struct CrossRepoPullRequestRequest: Sendable, Hashable {
-  let owner: String
-  let repo: String
-  let branches: [String]
-
-  var key: RepoKey {
-    RepoKey(owner: owner, repo: repo)
-  }
-}
-
-nonisolated struct CrossRepoPullRequestResult: Sendable {
-  let successByRepo: [RepoKey: [String: GithubPullRequest]]
-  let failedRepos: [RepoKey: GithubCLIError]
-
-  init(
-    successByRepo: [RepoKey: [String: GithubPullRequest]] = [:],
-    failedRepos: [RepoKey: GithubCLIError] = [:]
-  ) {
-    self.successByRepo = successByRepo
-    self.failedRepos = failedRepos
-  }
-}
-
 struct GithubCLIClient: Sendable {
   var defaultBranch: @Sendable (URL) async throws -> String
   var resolveRemoteInfo: @Sendable (URL) async -> GithubRemoteInfo?
@@ -128,105 +62,6 @@ extension DependencyValues {
   var githubCLI: GithubCLIClient {
     get { self[GithubCLIClient.self] }
     set { self[GithubCLIClient.self] = newValue }
-  }
-}
-
-private struct GithubPullRequestsRequest: Sendable {
-  let host: String
-  let owner: String
-  let repo: String
-}
-
-nonisolated private struct GithubRepoViewRemoteInfoResponse: Decodable, Sendable {
-  let owner: Owner
-  let name: String
-  let url: String
-
-  nonisolated var remoteInfo: GithubRemoteInfo? {
-    guard let host = URL(string: url)?.host else {
-      return nil
-    }
-    return GithubRemoteInfo(host: host, owner: owner.login, repo: name)
-  }
-
-  nonisolated struct Owner: Decodable, Sendable {
-    let login: String
-  }
-}
-
-private actor GithubCLIExecutableResolver {
-  private var cachedExecutableURL: URL?
-  private var inFlightResolution: Task<URL, Error>?
-
-  func executableURL(shell: ShellClient) async throws -> URL {
-    if let cachedExecutableURL {
-      return cachedExecutableURL
-    }
-    if let inFlightResolution {
-      return try await inFlightResolution.value
-    }
-    let resolutionTask = Task {
-      try await resolveExecutableURL(shell: shell)
-    }
-    inFlightResolution = resolutionTask
-    do {
-      let executableURL = try await resolutionTask.value
-      cachedExecutableURL = executableURL
-      inFlightResolution = nil
-      return executableURL
-    } catch {
-      inFlightResolution = nil
-      throw error
-    }
-  }
-
-  func invalidate() {
-    cachedExecutableURL = nil
-    inFlightResolution?.cancel()
-    inFlightResolution = nil
-  }
-
-  private func resolveExecutableURL(shell: ShellClient) async throws -> URL {
-    if let executableURL = await locateExecutableURL(
-      shell: shell,
-      useLoginShell: false
-    ) {
-      return executableURL
-    }
-    if let executableURL = await locateExecutableURL(
-      shell: shell,
-      useLoginShell: true
-    ) {
-      return executableURL
-    }
-    throw GithubCLIError.unavailable
-  }
-
-  private func locateExecutableURL(
-    shell: ShellClient,
-    useLoginShell: Bool
-  ) async -> URL? {
-    let whichURL = URL(fileURLWithPath: "/usr/bin/which")
-    do {
-      let output: String
-      if useLoginShell {
-        output = try await shell.runLogin(
-          whichURL,
-          ["gh"],
-          nil,
-          log: false
-        ).stdout
-      } else {
-        output = try await shell.run(whichURL, ["gh"], nil).stdout
-      }
-      let trimmed = output.trimmingCharacters(in: .whitespacesAndNewlines)
-      guard !trimmed.isEmpty else {
-        return nil
-      }
-      return URL(fileURLWithPath: trimmed)
-    } catch {
-      return nil
-    }
   }
 }
